requirements.txt
flask
google-generativeai
beautifulsoup4
requests
pypdf
markdown
